Made, hookCustomizer into a util, and converted hosts.js to use feathers pinia
